Removing unnecessary quotes xml files
