+++ features/quotes/src/main/java/com/mmk/quotes/presentation/di/ViewModelModule.kt
@@ -1,11 +1,11 @@
 package com.mmk.quotes.presentation.di
 
 import com.mmk.quotes.presentation.addnewquote.AddNewQuoteVM
-import com.mmk.quotes.presentation.allquotes.QuotesViewModel
+import com.mmk.quotes.presentation.allquotes.QuotesVM
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 internal val viewModelModule = module {
-    viewModel { QuotesViewModel(get()) }
+    viewModel { QuotesVM(get()) }
     viewModel { AddNewQuoteVM(get()) }
 }
