--- features/quotes/src/main/res/layout/fragment_quotes.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:theme="@style/AppTheme.Base"
-    android:layout_height="match_parent">
-
-
-    <TextView
-        android:id="@+id/quoteTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:fontFamily="@font/tthoves_bold"
-        android:text="@string/title_quotes"
-        android:textColor="?attr/colorOnBackground"
-        android:textSize="28sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/quotesRecyclerView"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginTop="16dp"
-        android:clipToPadding="false"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/quoteTitle"
-        tools:listitem="@layout/item_quote_list" />
-
-    <ProgressBar
-        android:id="@+id/progressBarQuotes"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="@id/quotesRecyclerView"
-        app:layout_constraintEnd_toEndOf="@id/quotesRecyclerView"
-        app:layout_constraintStart_toStartOf="@id/quotesRecyclerView"
-        app:layout_constraintTop_toTopOf="@id/quotesRecyclerView" />
-
-    <ImageView
-        android:id="@+id/emptyView"
-        android:layout_width="300dp"
-        android:layout_height="300dp"
-        android:src="@drawable/ic_empty_quotes"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
