--- features/quotes/src/main/res/layout/item_quote_list.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    app:cardBackgroundColor="?attr/colorSurface"
-    app:cardCornerRadius="16dp"
-    app:cardElevation="10dp"
-    android:theme="@style/AppTheme.Base"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_margin="16dp">
-
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageView
-            android:id="@+id/quoteIconImageView"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            android:layout_width="36dp"
-            android:layout_height="36dp"
-            android:layout_margin="20dp"
-            android:src="@drawable/ic_quote" />
-
-        <de.hdodenhof.circleimageview.CircleImageView
-            android:id="@+id/quoteAuthorImageView"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="1"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:src="@drawable/ic_person"
-            android:layout_width="80dp"
-            android:layout_height="80dp"
-            android:layout_margin="20dp"
-            android:scaleType="centerCrop" />
-
-        <TextView
-            android:id="@+id/quoteText"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/quoteAuthorImageView"
-            app:layout_constraintStart_toEndOf="@id/quoteIconImageView"
-            app:layout_constraintTop_toBottomOf="@id/quoteIconImageView"
-            app:layout_constraintVertical_bias="0"
-            tools:text="@string/sample_quote"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="24dp"
-            android:layout_marginEnd="24dp"
-            android:fontFamily="@font/tthoves_medium"
-            android:textColor="?attr/colorOnSurface"
-            android:textSize="20sp" />
-
-        <View
-            android:id="@+id/line_quote_author"
-            app:layout_constraintBottom_toBottomOf="@id/quoteAuthor"
-            app:layout_constraintStart_toStartOf="@id/quoteText"
-            app:layout_constraintTop_toTopOf="@id/quoteAuthor"
-            android:layout_width="16dp"
-            android:layout_height="1dp"
-            android:background="?attr/colorOnSurface" />
-
-        <TextView
-            android:id="@+id/quoteAuthor"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/quoteAuthorImageView"
-            app:layout_constraintHorizontal_bias="0"
-            app:layout_constraintStart_toEndOf="@id/line_quote_author"
-            app:layout_constraintTop_toBottomOf="@id/quoteText"
-            app:layout_constraintVertical_bias="0"
-            tools:text="@string/sample_author"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="20dp"
-            android:layout_marginBottom="20dp"
-            android:fontFamily="@font/tthoves_regular"
-            android:textColor="?attr/colorInactive"
-            android:textSize="16sp" />
-
-        <ImageView
-            android:id="@+id/quoteShareImageView"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/quoteAuthorImageView"
-            app:layout_constraintVertical_bias="1"
-            app:tint="?attr/colorOnSurface"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginBottom="20dp"
-            android:src="@drawable/ic_share" />
-
-        <ImageView
-            android:id="@+id/quoteLikeImageView"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/quoteShareImageView"
-            app:layout_constraintTop_toBottomOf="@id/quoteAuthorImageView"
-            app:layout_constraintVertical_bias="1"
-            app:tint="?attr/colorSecondary"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginBottom="20dp"
-            android:src="@drawable/ic_heart_unfilled" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-
-</com.google.android.material.card.MaterialCardView>
-
-
