--- features/quotes/src/main/res/layout/item_quote_load_state.xml
@@ -1,30 +1,0 @@
-<androidx.constraintlayout.widget.ConstraintLayout android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:paddingBottom="50dp"
-    android:orientation="vertical"
-    android:theme="@style/AppTheme.Base"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <ProgressBar
-        android:id="@+id/progressBar"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        style="?android:attr/progressBarStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:padding="8dp" />
-
-    <Button
-        android:id="@+id/buttonRetry"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="@string/btn_quotes_loading_retry"/>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
